refactor: 거래 페이지 컨테이너 분리 후 동적 import
